myproject/pyproject.toml
[build-system]
requires = [
  "setuptools >= 68",
  "reahl-component-metadata >= 7.0.0"]
build-backend = "setuptools.build_meta"
  
[project]
name = "myproject"
version = "2.0"
dependencies = [
  "project-b>=3.1,<3.2",
  "project-a>=1.4,<1.5",
  "project-c>=4.0,<4.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]

[tool.setuptools.packages.find]
exclude = ["etc", "build", "dist"]

[tool.reahl-component]
persisted = [
  "myproject.code:MyObject"
]

[tool.reahl-component.versions.'1.0']
dependencies = [
  "project-b>=2.0,<2.1",
  "project-a>=1.2,<1.3",
  "project-c>=3.0,<3.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]

migrations = [
  "myproject.migrations:CreateDB"
  ]


[tool.reahl-component.versions.'2.0']
migrations = [
  "myproject.migrations:ToNewC"
  ]
